requirements.txt
# API Clients
newsapi-python==0.2.7
anthropic==0.75.0
elevenlabs==1.53.0
requests==2.31.0
google-generativeai>=0.8.0

# Audio/Video Processing
ffmpeg-python==0.2.0
Pillow>=10.0.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2
PyJWT==2.8.0

# Logging
structlog==24.1.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
